Create a hand that is the same number of cards as hands in the game
